Rename metatag resolver to metatag helper
